requirements.txt
python-docx>=0.8.11
langchain>=0.0.267
openai>=0.27.8
anthropic>=0.3.11
pydantic>=2.0.0
numpy>=1.24.1
python-dotenv>=1.0.0
